# swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "todolist-app API",
    "version": "1.0.0",
    "description": "개인별 할일 관리 애플리케이션 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Access Token 재발급",
        "description": "Refresh Token으로 새 Access Token을 발급받습니다.",
        "operationId": "refreshToken",
        "description": "Refresh Token으로 새 Access Token을 발급받습니다. Refresh Token은 httpOnly Cookie에서 자동 전송됩니다.",
        "responses": {
          "200": {
            "description": "토큰 재발급 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                },
                "example": {
                  "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않은 Refresh Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INVALID_REFRESH_TOKEN",
                    "message": "유효하지 않거나 만료된 Refresh Token입니다"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/todos": {
      "post": {
        "tags": ["Todos"],
        "summary": "할일 생성",
        "description": "새 할일을 생성합니다.",
        "operationId": "createTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreate"
              },
              "example": {
                "title": "OpenAPI 스펙 작성",
                "description": "todolist-app의 Swagger 문서를 작성한다.",
                "start_date": "2026-04-01",
                "end_date": "2026-04-02"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "할일 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "유효성 검증에 실패했습니다"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Todos"],
        "summary": "할일 목록 조회",
        "description": "인증된 사용자의 할일 목록을 페이지네이션으로 조회합니다.",
        "operationId": "getTodos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "할일 상태 필터",
            "schema": {
              "$ref": "#/components/schemas/TodoStatus"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "description": "정렬 기준 (기본값: created_at)",
            "schema": {
              "type": "string",
              "enum": ["end_date", "start_date", "created_at", "title"],
              "default": "created_at"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "description": "정렬 방향 (기본값: desc)",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "페이지 당 항목 수 (기본값: 20)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TodoListItem"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  },
                  "required": ["data", "pagination"]
                },
                "example": {
                  "data": [
                    {
                      "id": "550e8400-e29b-41d4-a716-446655440001",
                      "title": "OpenAPI 스펙 작성",
                      "start_date": "2026-04-01",
                      "end_date": "2026-04-02",
                      "is_completed": false,
                      "completed_at": null,
                      "status": "IN_PROGRESS",
                      "created_at": "2026-04-01T00:00:00.000Z",
                      "updated_at": "2026-04-01T00:00:00.000Z"
                    }
                  ],
                  "pagination": {
                    "total": 42,
                    "page": 1,
                    "limit": 20,
                    "totalPages": 3
                  }
                }
              }
            }
          },
          "400": {
            "description": "유효하지 않은 status 또는 sort_by",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INVALID_QUERY_PARAM",
                    "message": "올바른 상태 값을 입력해주세요"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "할일 단건 조회",
        "description": "특정 할일의 전체 정보를 조회합니다.",
        "operationId": "getTodoById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할일 ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "할일 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 타 사용자 소유",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_NOT_FOUND",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Todos"],
        "summary": "할일 수정",
        "description": "특정 할일의 전체 내용을 수정합니다.",
        "operationId": "updateTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할일 ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdate"
              },
              "example": {
                "title": "OpenAPI 스펙 수정",
                "description": "todolist-app의 Swagger 문서를 업데이트한다.",
                "start_date": "2026-04-01",
                "end_date": "2026-04-03"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "할일 수정 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "유효성 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "유효성 검증에 실패했습니다"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 타 사용자 소유",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_NOT_FOUND",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "할일 삭제",
        "description": "특정 할일을 삭제합니다.",
        "operationId": "deleteTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할일 ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "할일 삭제 성공 (No Content)"
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 타 사용자 소유",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_NOT_FOUND",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{id}/complete": {
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 완료 처리",
        "description": "특정 할일을 완료 상태로 변경합니다. 완료 시각을 기준으로 기한 내 완료(COMPLETED_ON_TIME) 또는 지연 완료(COMPLETED_LATE) 상태가 결정됩니다.",
        "operationId": "completeTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할일 ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "완료 처리 성공 (전체 Todo 객체 반환)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 타 사용자 소유",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_NOT_FOUND",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          },
          "409": {
            "description": "이미 완료된 할일",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "ALREADY_COMPLETED",
                    "message": "이미 완료된 할일입니다"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{id}/uncomplete": {
      "patch": {
        "tags": ["Todos"],
        "summary": "할일 완료 취소",
        "description": "완료 상태의 할일을 미완료 상태로 되돌립니다. 현재 날짜를 기준으로 UPCOMING, IN_PROGRESS, OVERDUE 중 적절한 상태가 자동 결정됩니다.",
        "operationId": "uncompleteTodo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "할일 ID (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "완료 취소 성공 (전체 Todo 객체 반환)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "인증이 필요합니다"
                  }
                }
              }
            }
          },
          "404": {
            "description": "존재하지 않거나 타 사용자 소유",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "TODO_NOT_FOUND",
                    "message": "할일을 찾을 수 없습니다"
                  }
                }
              }
            }
          },
          "409": {
            "description": "완료되지 않은 할일",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_COMPLETED",
                    "message": "완료되지 않은 할일입니다"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token. Authorization: Bearer {accessToken}"
      }
    },
    "schemas": {
      "TodoStatus": {
        "type": "string",
        "enum": [
          "UPCOMING",
          "IN_PROGRESS",
          "OVERDUE",
          "COMPLETED_ON_TIME",
          "COMPLETED_LATE"
        ],
        "description": "할일 상태\n- UPCOMING: 시작일 이전\n- IN_PROGRESS: 진행 중 (시작일~종료일)\n- OVERDUE: 기한 초과 미완료\n- COMPLETED_ON_TIME: 기한 내 완료\n- COMPLETED_LATE: 기한 초과 후 완료"
      },
...
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "할일 고유 ID"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "소유자 사용자 ID"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "할일 제목"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 설명 (선택)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD)"
          },
          "is_completed": {
            "type": "boolean",
            "description": "완료 여부"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "완료 일시 (ISO8601, 미완료 시 null)"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시 (ISO8601)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "수정 일시 (ISO8601)"
          }
        },
        "required": [
          "id",
          "user_id",
          "title",
          "start_date",
          "end_date",
          "is_completed",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "TodoListItem": {
        "type": "object",
        "description": "할일 목록 조회 시 반환되는 항목 (description 미포함)",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "할일 고유 ID"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "할일 제목"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD)"
          },
          "is_completed": {
            "type": "boolean",
            "description": "완료 여부"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "완료 일시 (ISO8601, 미완료 시 null)"
          },
          "status": {
            "$ref": "#/components/schemas/TodoStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "생성 일시 (ISO8601)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "수정 일시 (ISO8601)"
          }
        },
        "required": [
          "id",
          "title",
          "start_date",
          "end_date",
          "is_completed",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "TodoCreate": {
        "type": "object",
        "description": "할일 생성 요청 본문",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "할일 제목"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 설명 (선택)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD)"
          }
        },
        "required": ["title", "start_date", "end_date"]
      },
      "TodoUpdate": {
        "type": "object",
        "description": "할일 수정 요청 본문",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "할일 제목"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true,
            "description": "할일 설명 (선택)"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "시작일 (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "종료일 (YYYY-MM-DD)"
          }
        },
        "required": ["title", "start_date", "end_date"]
      },
      "Pagination": {
        "type": "object",
        "description": "페이지네이션 메타 정보",
        "properties": {
          "total": {
            "type": "integer",
            "description": "전체 항목 수"
          },
          "page": {
            "type": "integer",
            "description": "현재 페이지 번호"
          },
          "limit": {
            "type": "integer",
            "description": "페이지 당 항목 수"
          },
          "totalPages": {
            "type": "integer",
            "description": "전체 페이지 수"
          }
        },
        "required": ["total", "page", "limit", "totalPages"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "공통 에러 응답 형식",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "에러 코드"
              },
              "message": {
                "type": "string",
                "description": "에러 메시지"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "새로 발급된 JWT Access Token (유효기간 30분)"
          }
        },
        "required": ["access_token"]
      },
...
    }
  }
}